feature:(views) views Installed react-router & react-router-dom Setup routes in App.js Implements Navbar, Getting started page & Examples page
