+++ package.json
@@ -45,6 +45,8 @@
     "prettier": "^1.14.3",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "serve": "^10.0.2",
     "webpack": "^4.35.0",
     "webpack-cli": "^3.1.2",
