+++ src/routes/GetStarted/index.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+const GetStarted = () => {
+  return (
+    <div style={{ width: '50%', margin: 'auto' }}>
+      <h1>The getting started landing page (docs page)</h1>
+      <div>
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
+          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
+          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in
+          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
+          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+          culpa qui officia deserunt mollit anim id est laborum.
+        </p>
+      </div>
+    </div>
+  );
+};
+
+export default GetStarted;
