+++ src/routes/Navbar/navbar.css
@@ -1,0 +1,28 @@
+.navbar-div {
+  display: flex;
+  width: 80%;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 auto;
+  background: #fff;
+  border-bottom: 1px solid #ddd;
+}
+
+ul {
+  display: inline-flex;
+  font-size: 23px;
+}
+
+a {
+  text-decoration: none;
+}
+
+li {
+  color: #000;
+  list-style-type: none;
+  padding: 15px;
+}
+
+li:hover {
+  color: gray;
+}
